set startDate as optional value when and added logic to get startDate in `WcStatsStore` if startDate is empty
